+++ errpref/zzzt_errpref_0002000_test.go
@@ -2,6 +2,118 @@ package errpref
 
 import "testing"
 
+func TestErrPrefixDto_CopyIn_000100(t *testing.T) {
+
+	initialStr := "Tx1.Something()\nTx2.SomethingElse()\nTx3.DoSomething()\nTx4() - Tx5()\nTx6.DoSomethingElse()\n"
+
+	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
+
+	// Setting Line Length to 60-Characters
+	ePDto.SetMaxErrPrefTextLineLength(60)
+
+	ePDto.SetEPrefCtx(
+		"Tx7.TrySomethingNew()",
+		"")
+
+	ePDto.SetCtx("something->newSomething")
+
+	ePDto.SetEPrefCtx(
+		"Tx8.TryAnyCombination()",
+		"")
+
+	ePDto.SetEPref("Tx9.TryAHammer()")
+
+	ePDto.SetCtx("x->y")
+
+	ePDto.SetEPref("Tx10.X()")
+
+	ePDto.SetEPrefCtx(
+		"Tx11.TryAnything()",
+		"")
+
+	ePDto.SetEPrefCtx(
+		"Tx12.TryASalad()",
+		"")
+
+	ePDto.SetEPref("Tx13.SomeFabulousAndComplexStuff()")
+
+	ePDto.SetEPrefCtx(
+		"Tx14.MoreAwesomeGoodness",
+		"A=7 B=8 C=9")
+
+	ePDto2 := ErrPrefixDto{}.Ptr()
+
+	err := ePDto2.CopyIn(
+		&ePDto,
+		"TestErrPrefixDto_CopyIn_000100")
+
+	if err != nil {
+		t.Errorf("Error:\n"+
+			"%v", err.Error())
+		return
+	}
+
+	if !ePDto.Equal(ePDto2) {
+		t.Error("Expected ePDto to Equal ePDto2\n" +
+			"However, THEY ARE NOT EQUAL!\n")
+	}
+}
+
+func TestErrPrefixDto_CopyOut_000100(t *testing.T) {
+
+	initialStr := "Tx1.Something()\nTx2.SomethingElse()\nTx3.DoSomething()\nTx4() - Tx5()\nTx6.DoSomethingElse()\n"
+
+	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
+
+	// Setting Line Length to 60-Characters
+	ePDto.SetMaxErrPrefTextLineLength(60)
+
+	ePDto.SetEPrefCtx(
+		"Tx7.TrySomethingNew()",
+		"")
+
+	ePDto.SetCtx("something->newSomething")
+
+	ePDto.SetEPrefCtx(
+		"Tx8.TryAnyCombination()",
+		"")
+
+	ePDto.SetEPref("Tx9.TryAHammer()")
+
+	ePDto.SetCtx("x->y")
+
+	ePDto.SetEPref("Tx10.X()")
+
+	ePDto.SetEPrefCtx(
+		"Tx11.TryAnything()",
+		"")
+
+	ePDto.SetEPrefCtx(
+		"Tx12.TryASalad()",
+		"")
+
+	ePDto.SetEPref("Tx13.SomeFabulousAndComplexStuff()")
+
+	ePDto.SetEPrefCtx(
+		"Tx14.MoreAwesomeGoodness",
+		"A=7 B=8 C=9")
+
+	ePDto2,
+		err := ePDto.CopyOut("TestErrPrefixDto_CopyOut_000100")
+
+	if err != nil {
+		t.Errorf("Error:\n"+
+			"%v", err.Error())
+		return
+	}
+
+	if !ePDto.Equal(&ePDto2) {
+		t.Error("Expected ePDto to Equal ePDto2\n" +
+			"However, THEY ARE NOT EQUAL!\n")
+	}
+
+}
+
 func TestErrPrefixDto_Multiple_000100(t *testing.T) {
 
 	initialStr := "Tx1.Something()\nTx2.SomethingElse()\nTx3.DoSomething()\nTx4() - Tx5()\nTx6.DoSomethingElse()\n"
@@ -1008,4 +1120,68 @@ func TestErrPrefixDto_String_000200(t *testing.T) {
 			actualStr)
 	}
 
+}
+
+func TestErrPrefixDto_StrMaxLineLen_000200(t *testing.T) {
+
+	initialStr := "Tx1.Something()\nTx2.SomethingElse()\nTx3.DoSomething()\nTx4() - Tx5()\nTx6.DoSomethingElse()\n"
+
+	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
+
+	ePDto.SetEPrefCtx(
+		"Tx7.TrySomethingNew()",
+		"")
+
+	ePDto.SetCtx("something->newSomething")
+
+	ePDto.SetEPrefCtx(
+		"Tx8.TryAnyCombination()",
+		"")
+
+	ePDto.SetEPref("Tx9.TryAHammer()")
+
+	ePDto.SetCtx("x->y")
+
+	ePDto.SetEPref("Tx10.X()")
+
+	ePDto.SetEPrefCtx(
+		"Tx11.TryAnything()",
+		"")
+
+	ePDto.SetEPrefCtx(
+		"Tx12.TryASalad()",
+		"")
+
+	ePDto.SetEPref("Tx13.SomeFabulousAndComplexStuff()")
+
+	ePDto.SetEPrefCtx(
+		"Tx14.MoreAwesomeGoodness",
+		"A=7 B=8 C=9")
+
+	expectedStr :=
+		"Tx1.Something() - Tx2.SomethingElse() - Tx3.DoSomething()\\n" +
+			"Tx4() - Tx5() - Tx6.DoSomethingElse()\\n" +
+			"Tx7.TrySomethingNew() : something->newSomething\\n" +
+			"Tx8.TryAnyCombination() - Tx9.TryAHammer() : x->y - Tx10.X()\\n" +
+			"Tx11.TryAnything() - Tx12.TryASalad()\\n" +
+			"Tx13.SomeFabulousAndComplexStuff()\\n" +
+			"Tx14.MoreAwesomeGoodness : A=7 B=8 C=9"
+
+	// Setting Line Length to 60-Characters
+	actualStr := ePDto.StrMaxLineLen(60)
+
+	actualStr = ErrPref{}.ConvertNonPrintableChars(
+		[]rune(actualStr),
+		false)
+
+	if expectedStr != actualStr {
+
+		t.Errorf("Error:"+
+			"Expected actualStr= '%v'\n"+
+			"Instead, actualStr= '%v'\n",
+			expectedStr,
+			actualStr)
+	}
+
+	ePDto.SetMaxErrPrefTextLineLengthToDefault()
 }